Rearranged util methods and removed unused code
